refactor: split the last two files over the 500-LoC cap
The merge with main pushed two more files past the cap (they were under
it when I checked pre-merge). Same treatment, exact piecewise move:

  ops/ssm_mamba.rs              541 -> 407  (+ ops/ssm_ssd.rs: the SSD
                                             chunked-scan launchers and
                                             their SSD_L/SSD_PT constants)
  weight_loader/nemotron.rs     504 -> 403  (+ nemotron/ssm_layer.rs: the
                                             LinearAttention arm, now
                                             Self::build_ssm_layer)

Gates re-run on the merged tree: fmt, clippy --workspace --tests, docs,
cargo test --workspace --locked, and the 500-LoC cap replayed with the
workflow's own allow-list -- 0 violations. GPU re-verified after the
loader restructure (weight loading is on that path): 1k TTFT 414ms,
decode 33.2 tok/s, coherence probes clean.
